download/pyproject.toml
[project]
name = "fakeShop"
version = "0.1.0"
description = "None"
readme = "README.md"
authors = [{ name = "AllDotPy", email = "hello@alldotpy.com" }]
maintainers = [{ name = "#Einswilli", email = "einswilligoeh@email.com" }]
requires-python = ">=3.12"
dependencies = [
    "fletxr==0.1.4.b1",
    "flet[all]>=0.28.3",
    "flet-webview==0.1.0",
    "httpx>=0.28.1",
    "fletcarousel>=0.0.14",
]
